html report capabilities
