+++ .ochrona.yml
@@ -4,10 +4,10 @@
 # dir: .
 # report_type: JSON
 # report_location: .
+# ignore: requests
 # include_dev: false
 # color_output: false
 # policies:
-#  - policy_type: license_type
-#    deny_list: APSL,GPL-PA,JSON
-#  - policy_type: package_name
-#    deny_list: jsont
+#  - license_type NIN APSL,GPL-PA,JSON
+# sbom: true
+# sbom_format: JSON
