+++ README.md
@@ -27,8 +27,7 @@
       - [Allowed Logical Operators](#allowed-logical-operators)
       - [Special Values](#special-values)
       - [Policy Examples](#policy-examples)
-    + [Legacy Policies](#legacy-policies)
-      - [Legacy Policy Types](#legacy-policy-types)
+- [SBOM](#sbom)
 - [Usage Examples](#usage-examples)
     + [Default Mode](#default-mode)
     + [Standard error code with Junit XML reporting saved to file](#standard-error-code-with-junit-xml-reporting-saved-to-file)
