+++ ochrona/reporter/reporter.py
@@ -4,17 +4,8 @@ Ochrona-cli
 :author: ascott
 """
 
-from dataclasses import asdict
-import datetime
-import json
-import os
-import re
-import shutil
 import sys
-import textwrap
-from typing import Any, Dict, List, Optional
-import xml.etree.ElementTree as ET
-from xml.dom import minidom
+from typing import List
 
 from rich import box, print
 from rich.table import Table
@@ -22,8 +13,13 @@ from rich.table import Table
 from ochrona.config import OchronaConfig
 from ochrona.log import OchronaLogger
 from ochrona.model.confirmed_vulnerability import Vulnerability
-from ochrona.model.policy_violation import PolicyViolation
 from ochrona.model.dependency_set import DependencySet
+from ochrona.reporter.reports.base import BaseReport
+from ochrona.reporter.reports.basic import BasicReport
+from ochrona.reporter.reports.full import FullReport
+from ochrona.reporter.reports.html import HTMLReport
+from ochrona.reporter.reports.json import JSONReport
+from ochrona.reporter.reports.xml import XMLReport
 
 
 class OchronaReporter:
@@ -75,51 +71,45 @@ class OchronaReporter:
         :return:
         """
         if self._report_type == "BASIC":
-            BaseReport.print_report_number(index, total, self._config.color_output)
-            BasicReport.print_findings(
-                vulnerabilities=result.confirmed_vulnerabilities,
-                violations=result.policy_violations,
+            BasicReport.generate(
+                result=result,
+                config=self._config,
                 source=source,
-                color=self._config.color_output,
+                total=total,
+                index=index,
             )
         elif self._report_type == "FULL":
-            BaseReport.print_report_number(index, total, self._config.color_output)
-            FullReport().print_findings(
-                vulnerabilities=result.confirmed_vulnerabilities,
-                violations=result.policy_violations,
+            FullReport().generate(
+                result=result,
+                config=self._config,
                 source=source,
-                color=self._config.color_output,
+                total=total,
+                index=index,
             )
         elif self._report_type == "JSON":
-            report = result.confirmed_vulnerabilities
-            violations = result.policy_violations
-
-            if report or violations:
-                if not self._report_location:
-                    BaseReport.print_report_number(
-                        index, total, self._config.color_output
-                    )
-                    JSONReport.display_report(report, violations, source, total, index)
-                else:
-                    JSONReport.save_report_to_file(
-                        report, violations, self._report_location, source, index
-                    )
-            elif (not report and not violations) and not self._report_location:
-                BaseReport.print_report_number(index, total, self._config.color_output)
-                JSONReport.display_report(report, violations, source, total, index)
-            else:
-                JSONReport.save_report_to_file(
-                    report, violations, self._report_location, source, index
-                )
-
+            JSONReport.generate(
+                result=result,
+                config=self._config,
+                source=source,
+                total=total,
+                index=index,
+            )
         elif self._report_type == "XML":
-            if not self._report_location:
-                BaseReport.print_report_number(index, total, self._config.color_output)
-                XMLReport.display_report(result, source, total, index)
-            else:
-                XMLReport.save_report_to_file(
-                    result, self._report_location, source, index
-                )
+            XMLReport.generate(
+                result=result,
+                config=self._config,
+                source=source,
+                total=total,
+                index=index,
+            )
+        elif self._report_type == "HTML":
+            HTMLReport.generate(
+                result=result,
+                config=self._config,
+                source=source,
+                total=total,
+                index=index,
+            )
 
     def _filter_ignored_vuln(self, confirmed_vuln: Vulnerability) -> bool:
         """
@@ -133,282 +123,3 @@ class OchronaReporter:
             if cv == confirmed_vuln.cve_id or cv == confirmed_vuln.name:
                 return False
         return True
-
-
-class BaseReport:
-    INFO = "[blue]"
-    ENDC = "[/]"
-    NO = "default"
-
-    @staticmethod
-    def print_report_number(index: int, total: int, color: bool = True):
-        print(
-            f"{os.linesep}{BaseReport.INFO if color else BaseReport.NO}Report {index + 1} of {total}{BaseReport.ENDC if color else BaseReport.NO}"
-        )
-
-    @staticmethod
-    def print_report_source(source: str, color: bool = True):
-        if color:
-            print(f"[bold white italics]File: {source}[/]")
-        else:
-            print(f"Analysis: {source}")
-
-
-class BasicReport(BaseReport):
-    """
-    Basic Report
-        - Includes package name, current version, affected version(s), CPE, Severity
-
-    This report can be logged to stdout
-    """
-
-    @staticmethod
-    def print_findings(
-        vulnerabilities: List[Vulnerability],
-        violations: List[PolicyViolation],
-        source: str,
-        color: bool = True,
-    ):
-        term_size = shutil.get_terminal_size((100, 20))
-        print(f"[bold white italics]File: {source}[/]")
-        table = Table(
-            box=box.ROUNDED, header_style="bold white", min_width=term_size.columns
-        )
-        table.add_column(
-            header="Vulnerability Check", style="blue bold", justify="right", width=30
-        )
-        table.add_column(
-            header=f"{'[bold red]:cross_mark: {} Vulnerabilities Detected![/]'.format(len(vulnerabilities)) if len(vulnerabilities) > 0 else '[bold green]:white_heavy_check_mark: No Vulnerabilities Detected![/]'}",
-            style="blue",
-            width=term_size.columns - 30,
-        )
-        if len(vulnerabilities) > 0:
-            for finding in vulnerabilities:
-                table.add_row("Package Name", finding.name)
-                table.add_row("Installed Version", finding.found_version)
-                table.add_row("CVE/Vuln ID", finding.cve_id)
-                table.add_row("Severity", f"{finding.ochrona_severity_score}")
-                if len(finding.affected_versions) > 0:
-                    affected_versions = os.linesep.join(
-                        [
-                            f"{f.get('operator', '')}{f.get('version_value', '')}"  # type: ignore
-                            for f in finding.affected_versions
-                        ]
-                    )
-                    table.add_row(
-                        "Affect Versions", affected_versions, end_section=True
-                    )
-                else:
-                    table.add_row(
-                        "Affect Versions",
-                        finding.vulnerable_version_expression.replace("version", ""),  # type: ignore
-                        end_section=True,
-                    )
-        table.add_row("", "")
-        table.add_row(
-            "[bold white] Policy Check[/]",
-            f"{'[bold red]:cross_mark: {} Policy Violations Found![/]'.format(len(violations)) if len(violations) > 0 else '[bold green]:white_heavy_check_mark: No Policy Violations Found![/]'}",
-        )
-        for violation in violations:
-            table.add_row("Policy", violation.friendly_policy_type)
-            table.add_row("Violation", violation.message)
-        print(table)
-
-
-class FullReport(BaseReport):
-    """
-    Full Report
-        - Includes package name, current version, affected version(s), CPE,
-                 Severity, Description, References, Publish Date
-
-    This report can be logged to stdout
-    """
-
-    @staticmethod
-    def print_findings(
-        vulnerabilities: List[Vulnerability],
-        violations: List[PolicyViolation],
-        source: str,
-        color: bool = True,
-    ):
-        term_size = shutil.get_terminal_size((100, 20))
-        print(f"[bold white italics]File: {source}[/]")
-        table = Table(
-            box=box.ROUNDED, header_style="bold white", min_width=term_size.columns
-        )
-        table.add_column(
-            header="Vulnerability Check", style="blue bold", justify="right", width=30
-        )
-        table.add_column(
-            header=f"{'[bold red]:cross_mark: {} Vulnerabilities Detected![/]'.format(len(vulnerabilities)) if len(vulnerabilities) > 0 else '[bold green]:white_heavy_check_mark: No Vulnerabilities Detected![/]'}",
-            style="blue",
-            width=term_size.columns - 30,
-        )
-        if len(vulnerabilities) > 0:
-            for finding in vulnerabilities:
-                table.add_row("Package Name", finding.name)
-                table.add_row("Installed Version", finding.found_version)
-                table.add_row("Reason", finding.reason)
-                table.add_row("CVE/Vuln ID", finding.cve_id)
-                table.add_row("Vulnerability Publish Date", finding.publish_date)
-                table.add_row("Severity", f"{finding.ochrona_severity_score}")
-                table.add_row("Description", finding.description)
-                if len(finding.affected_versions) > 0:
-                    affected_versions = os.linesep.join(
-                        [
-                            f"{f.get('operator', '')}{f.get('version_value', '')}"  # type: ignore
-                            for f in finding.affected_versions
-                        ]
-                    )
-                    table.add_row("Affect Versions", affected_versions)
-                else:
-                    table.add_row(
-                        "Affect Versions",
-                        finding.vulnerable_version_expression.replace("version", ""),  # type: ignore
-                    )
-                table.add_row("License", finding.license)
-                references = os.linesep.join(
-                    [f"[magenta underline]{ref}[/]" for ref in finding.references]
-                )
-                table.add_row("References", references, end_section=True)
-        table.add_row("", "")
-        table.add_row(
-            "[bold white] Policy Check[/]",
-            f"{'[bold red]:cross_mark: {} Policy Violations Found![/]'.format(len(violations)) if len(violations) > 0 else '[bold green]:white_heavy_check_mark: No Policy Violations Found![/]'}",
-        )
-        for violation in violations:
-            table.add_row("Policy", violation.friendly_policy_type)
-            table.add_row("Violation", violation.message)
-        print(table)
-
-
-class JSONReport(BaseReport):
-    """
-    JSON report
-        - Includes full API findings and metadata in json format.
-
-    This report can be logged to stdout
-    """
-
-    @staticmethod
-    def display_report(
-        result: List[Vulnerability],
-        violations: List[PolicyViolation],
-        source: str,
-        total: int,
-        index: int,
-    ):
-        BaseReport.print_report_source(source)
-        print(JSONReport.generate_report_body(result, violations, source))
-
-    @staticmethod
-    def save_report_to_file(
-        result: List[Vulnerability],
-        violations: List[PolicyViolation],
-        location: str,
-        source: str,
-        index: int,
-    ):
-        with open(
-            JSONReport.generate_report_filename(location, source, index), "w"
-        ) as f:
-            f.write(JSONReport.generate_report_body(result, violations, source))
-
-    @staticmethod
-    def generate_report_body(
-        result: List[Vulnerability], violations: List[PolicyViolation], source: str
-    ) -> str:
-        report = {
-            "meta": {
-                "source": str(source),
-                "timestamp": datetime.datetime.now().isoformat(),
-            },
-            "findings": [asdict(r) for r in result],
-            "violations": [asdict(v) for v in violations],
-        }
-        return json.dumps(report, indent=4)
-
-    @staticmethod
-    def generate_report_filename(location: str, source: str, index: int) -> str:
-        return f"{location}/{index+1}_{os.path.basename(source).lower()}_results.json"
-
-
-class XMLReport(BaseReport):
-    """
-    XML report
-        - Includes only discovered vulnerabilities
-    This report can be logged to stdout
-    """
-
-    VIOLATION_PACKAGE_PATTERN = r"^.*\.\s+\(from (.*)\)$"
-
-    @staticmethod
-    def display_report(result: DependencySet, source: str, total: int, index: int):
-        BaseReport.print_report_source(source)
-        print(XMLReport.generate_report_body(result, source))
-
-    @staticmethod
-    def save_report_to_file(
-        result: DependencySet, location: str, source: str, index: int
-    ):
-        with open(
-            XMLReport.generate_report_filename(location, source, index), "w"
-        ) as f:
-            f.write(XMLReport.generate_report_body(result, source))
-
-    @staticmethod
-    def generate_report_body(result: DependencySet, source: str) -> str:
-        suites = ET.Element("testsuites")
-        suite = ET.SubElement(suites, "testsuite")
-        suite.set("tests", str(len(result.flat_list)))
-        props = ET.SubElement(suite, "properties")
-        source_prop = ET.SubElement(props, "property")
-        source_prop.set("name", "source")
-        source_prop.set("value", str(source))
-        ts_prop = ET.SubElement(props, "property")
-        ts_prop.set("name", "timestamp")
-        ts_prop.set("value", datetime.datetime.now().isoformat())
-        for dep in result.flat_list:
-            case = ET.SubElement(suite, "testcase")
-            case.set("classname", "ochronaDependencyVulnCheck")
-            case.set("name", dep)
-        if len(result.confirmed_vulnerabilities) > 0:
-            for vuln in result.confirmed_vulnerabilities:
-                case = list(
-                    filter(
-                        lambda x: x.get("name") == vuln.found_version,
-                        list(suite.iter()),
-                    )
-                )[0]
-                failure = ET.SubElement(case, "failure")
-                failure.set("type", "confirmed_vulnerability")
-                failure.text = f"Package name: {vuln.name}\nVulnerability description: {vuln.description}\nCVE: {vuln.cve_id}\nSeverity: {vuln.ochrona_severity_score}"
-        if len(result.policy_violations) > 0:
-            for dep in result.flat_list:
-                case = ET.SubElement(suite, "testcase")
-                case.set("classname", "ochronaDependencyPolicyCheck")
-                case.set("name", dep)
-            for violation in result.policy_violations:
-                violating_package = re.search(  # type: ignore
-                    XMLReport.VIOLATION_PACKAGE_PATTERN, violation.message
-                ).groups(1)[0]
-                case = list(
-                    filter(
-                        lambda x: x.get("name") == violating_package
-                        and x.get("classname") == "ochronaDependencyPolicyCheck",
-                        list(suite.iter()),
-                    )
-                )[0]
-                failure = ET.SubElement(case, "failure")
-                failure.set("type", "policy_violations")
-                failure.text = violation.message
-        return minidom.parseString(ET.tostring(suites)).toprettyxml(indent="   ")
-
-    @staticmethod
-    def generate_report_filename(location: str, source: str, index: int) -> str:
-        return f"{location}/{index + 1}_{os.path.basename(source).lower()}_results.xml"
-
-
-class HTMLReport(BaseReport):
-    # TODO
-    pass
