+++ ochrona/schema/spdx_modified.json
@@ -5889,7 +5889,8 @@
             "aliases": [
                 "PSF",
                 "PSF license",
-                "PSF-2.0"
+                "PSF-2.0",
+                "Python Software Foundation License"
             ]
         },
         {
