+++ requirements.txt
@@ -15,3 +15,5 @@ rich==10.16.1
 Pygments==2.11.1
 commonmark==0.9.1
 colorama==0.4.4
+Jinja2==3.0.3
+MarkupSafe==2.0.1
